Remove `GetsOutputHTML` trait and restore `SeoProTags`.
